feat: add no-telemetry feature flag for minimal builds without OpenTelemetry
- Add conditional compilation for telemetry-grpc (default), telemetry-http, and no-telemetry features
- Introduce OtelRecorder wrapper around ScottyMetrics with clean API
- Add NoOpRecorder with zero-cost inlined methods for no-telemetry builds
- Refactor metrics call sites to use metrics() function instead of deprecated get_metrics()
- Remove backwards compatibility modules (shell.rs, websocket.rs, tasks.rs)
- Add feature conflict checks in main.rs
- Preserve all detailed metrics functionality (app list, tokio runtime, memory, etc)
- Support simultaneous telemetry-grpc + telemetry-http features (runtime config)

+++ scotty/src/tasks/output_streaming.rs
@@ -139,7 +139,7 @@ impl TaskOutputStreamingService {
                                     }),
                                 )
                                 .await;
-                            crate::metrics::tasks::record_output_lines(chunk_len);
+                            crate::metrics::tasks::record_output_lines(chunk_len as usize);
                         }
 
                         // Update last sent sequence to the last historical line
@@ -212,7 +212,7 @@ impl TaskOutputStreamingService {
                                         has_more: false,
                                     }),
                                 ).await;
-                                crate::metrics::tasks::record_output_lines(lines_count as u64);
+                                crate::metrics::tasks::record_output_lines(lines_count);
                             }
                         }
                     }
@@ -272,7 +272,7 @@ impl TaskOutputStreamingService {
                                                 has_more: false,
                                             }),
                                         ).await;
-                                        crate::metrics::tasks::record_output_lines(lines_count);
+                                crate::metrics::tasks::record_output_lines(lines_count as usize);
                                     }
                                 }
                             }
@@ -297,7 +297,7 @@ impl TaskOutputStreamingService {
                         }),
                     )
                     .await;
-                crate::metrics::tasks::record_output_lines(lines_count);
+                crate::metrics::tasks::record_output_lines(lines_count as usize);
             }
 
             // Send stream ended message
